requirements.txt
numpy>=1.24
pandas>=1.5
scikit-learn>=1.2
imbalanced-learn>=0.10.1
xgboost>=1.7
lightgbm>=4.0
catboost>=1.1
matplotlib>=3.6
seaborn>=0.12
shap>=0.42
optuna>=3.0
joblib
torch>=2.0
torchvision
tensorboard
Pillow
opencv-python
pytest
tqdm
